Move contents of cli-commands/signing to sidechain-domain/crypto
