+++ toolkit/cli/commands/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod address_association_signatures;
 pub mod key_params;
 pub mod registration_signatures;
-pub mod signing;
