+++ toolkit/partner-chains-cli/src/register/register1.rs
@@ -6,13 +6,13 @@ use crate::keystore::CROSS_CHAIN;
 use crate::{config::config_fields, *};
 use anyhow::anyhow;
 use cli_commands::registration_signatures::RegisterValidatorMessage;
-use cli_commands::signing::sc_public_key_and_signature_for_datum;
 use ogmios::config::prompt_ogmios_configuration;
 use ogmios::get_shelley_config;
 use partner_chains_cardano_offchain::csl::NetworkTypeExt;
 use select_utxo::{query_utxos, select_from_utxos};
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
+use sidechain_domain::crypto::sc_public_key_and_signature_for_datum;
 use sidechain_domain::{NetworkType, SidechainPublicKey, UtxoId};
 use sp_core::bytes::from_hex;
 use sp_core::{ecdsa, Pair};
